+++ src/Mapper/Complete.php
@@ -2,6 +2,8 @@
 
 namespace FiiSoft\Jackdaw\Mapper;
 
+use FiiSoft\Jackdaw\Internal\Helper;
+
 final class Complete implements Mapper
 {
     /** @var Mapper */
@@ -16,7 +18,7 @@ final class Complete implements Mapper
      */
     public function __construct($field, $mapper)
     {
-        if ($this->isFieldValid($field)) {
+        if (Helper::isFieldValid($field)) {
             $this->field = $field;
         } else {
             throw new \InvalidArgumentException('Invalid param field');
@@ -39,10 +41,5 @@ final class Complete implements Mapper
             $key => $value,
             $this->field => $this->mapper->map($value, $key),
         ];
-    }
-    
-    private function isFieldValid($field): bool
-    {
-        return \is_string($field) && $field !== '' || \is_int($field);
     }
 }