+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All important changes to `fiisoft-jackdaw-stream` will be documented in this fil
 ## 1.2.0 - incompatible changes!
 
 - method Reducer::consume changed, key is passed as second argument
+- added method StreamApi::moveTo
 
 ## 1.1.0 - incompatible changes!